+++ src/popup/pages/ShowPrivateKey/index.js
@@ -7,24 +7,40 @@ import CustomView from "../../component/CustomView";
 import Toast from "../../component/Toast";
 import "./index.scss";
 import copy_address from "../../../assets/images/copy_address.png";
+import SecurityPwd from "../../component/SecurityPwd";
+import { sendMsg } from "../../../utils/commonMsg";
+import { WALLET_GET_PRIVATE_KEY } from "../../../constant/types";
+import { SEC_SHOW_PRIVATE_KEY } from "../../../constant/walletType";
 class ShowPrivateKeyPage extends React.Component {
   constructor(props) {
     super(props);
-    let privateKey = props.location.params?.privateKey ?? ""
-    let address = props.location.params?.address ?? ""
     this.state = {
-      priKey: privateKey,
-      address,
+      priKey: "",
+      showSecurity:true
     };
+    this.address = props.location.params?.address ?? ""
+    this.isUnMounted = false;
   }
 
+  componentWillUnmount() {
+    this.isUnMounted = true;
+  }
+  callSetState = (data, callback) => {
+    if (!this.isUnMounted) {
+      this.setState({
+        ...data
+      }, () => {
+        callback && callback()
+      })
+    }
+  }
   renderTip = () => {
     return (
       <p className="wallet-tip-description">{getLanguage('privateKeyTip_2')}</p>
     )
   }
   copyAddress = () => {
-    copyText(this.state.address).then(() => {
+    copyText(this.address).then(() => {
       Toast.info(getLanguage('copySuccess'))
     })
   }
@@ -32,7 +48,7 @@ class ShowPrivateKeyPage extends React.Component {
     return (<div className={"click-cursor"} onClick={this.copyAddress}>
       <p className="wallet-show-title">{getLanguage('walletAddress')}</p>
       <div className={"wallet-pri-address-container"}>
-      <p className="wallet-show-address">{this.state.address}</p>
+      <p className="wallet-show-address">{this.address}</p>
       </div>
     </div>)
   }
@@ -70,20 +86,45 @@ class ShowPrivateKeyPage extends React.Component {
       </div>
     )
   }
-  goToNext = () => {
-    this.props.history.goBack()
-  }
 
+  onClickCheck=(password)=>{
+    sendMsg({
+      action: WALLET_GET_PRIVATE_KEY,
+      payload: {
+        password: password,
+        address: this.address
+      }
+    },
+      async (privateKey) => {
+        if (privateKey.error) {
+          if (privateKey.type === "local") {
+            Toast.info(getLanguage(privateKey.error))
+          } else {
+            Toast.info(privateKey.error)
+          }
+        } else {
+          this.callSetState({
+            priKey:privateKey,
+            showSecurity:false
+          },()=>{
+            Toast.info(getLanguage("securitySuccess"))
+          })
+        }
+      })
+  }
   render() {
+    const {showSecurity} = this.state
+    let title = showSecurity? getLanguage('securityPassword'):getLanguage('showPrivateKey')
     return (
       <CustomView
-        title={getLanguage('showPrivateKey')}
+        title={title}
         history={this.props.history}>
+        {showSecurity ? <SecurityPwd onClickCheck={this.onClickCheck} action={SEC_SHOW_PRIVATE_KEY}/>:
         <div className="mne-show-container">
           {this.renderAddress()}
           {this.renderKey()}
           {this.renderCopy()}
-        </div>
+        </div>}
       </CustomView>)
   }
 }
